Изменение в .gitignore и README.md. Изменения типа цен с int на float

+++ migrations/m231105_183228_CreateProductsTable.php
@@ -16,7 +16,7 @@ class m231105_183228_CreateProductsTable extends Migration
             'id' => $this->primaryKey(),
             'name' => $this->string()->notNull(),
             'description' => $this->text()->null(),
-            'price' => $this->integer()->notNull(),
+            'price' => $this->float()->notNull(),
         ]);
     }
